refactor: split test_create.py up
